autoconfiscated bertini_real.  intended to be built against autoconfiscated bertini, which at the time of this writing is to be released.
