+++ documentation/bertini_real.doxy.config
@@ -755,7 +755,8 @@ WARN_LOGFILE           =
 
 INPUT                  = ../include \
                          ../src \
-                         ./
+                         ./ \
+                         README.md
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
@@ -893,7 +894,7 @@ FILTER_SOURCE_PATTERNS =
 # (index.html). This can be useful if you have a project on for instance GitHub
 # and want to reuse the introduction page also for the doxygen output.
 
-USE_MDFILE_AS_MAINPAGE =
+USE_MDFILE_AS_MAINPAGE = YES
 
 #---------------------------------------------------------------------------
 # Configuration options related to source browsing
