+++ include/programConfiguration.hpp
@@ -24,7 +24,7 @@
 
 #include "bertini_extensions.hpp"
 
-#define BERTINI_REAL_VERSION_STRING "0.1.0"
+#define BERTINI_REAL_VERSION_STRING "0.1.1"
 #define SAMPLER_VERSION_STRING "0.9.9"
