+++ matlab_codes/@bertini_real_plotter/bertini_real_plotter.m
@@ -57,9 +57,6 @@ classdef bertini_real_plotter < handle
 		
 		is_bounded = [];
 		fv = [];
-				
-		format = '';
-		format_flag = '';
 	end
 	
 	methods
@@ -88,8 +85,8 @@ classdef bertini_real_plotter < handle
 		function set_default_options(br_plotter)
 			br_plotter.options.use_custom_projection = false;
 			br_plotter.options.markersize = 10;
-			br_plotter.options.sample_alpha = 0.5;
-			br_plotter.options.face_alpha = 0.5;
+			br_plotter.options.sample_alpha = 1;
+			br_plotter.options.face_alpha = 1;
 			br_plotter.options.edge_alpha = 0.4;
 			br_plotter.options.fontsizes.legend = 12;
 			br_plotter.options.fontsizes.labels = 16;
@@ -405,7 +402,8 @@ classdef bertini_real_plotter < handle
 			if br_plotter.options.autosave
 				try
 					save_routine(br_plotter);
-				catch
+				catch exception
+					display(exception);
 					display('saving render not completed');
 				end
 			end
@@ -511,7 +509,6 @@ classdef bertini_real_plotter < handle
 		center_camera_on_selected_point(br_plotter,source, event)
 		
 		save_routine(br_plotter,varargin) % is a callback function
-		render_into_file(br_plotter)
 		
 		
 		% setup the interactive buttons
