+++ matlab_codes/@bertini_real_plotter/set_initial_visibility.m
@@ -60,7 +60,7 @@ switch br_plotter.dimension
 			br_plotter.switches.display_faces = 0;
 			br_plotter.switches.display_face_samples = 1;
 			
-			br_plotter.switches.curve_refinements.main = 1;
+			br_plotter.switches.curve_refinements.main = 0;
 			br_plotter.switches.curve_refinements.show_midslices = 0;
 			br_plotter.switches.curve_refinements.show_critslices = 0;
 			br_plotter.switches.curve_refinements.show_spherecurve = 1;
