+++ matlab_codes/@bertini_real_plotter/setupfig.m
@@ -6,6 +6,7 @@ function setupfig(br_plotter,varargin)
 
 	if ~isempty(which('getmondim'))
 		mondim = getmondim;
+		mondim(3:4) = min(mondim(3:4));
 		br_plotter.window = [20 20 mondim(3)-60 mondim(4)-130];
 	end
