+++ src/body/checkSelfConjugate.cpp
