+++ src/body/curve.cpp
