+++ src/solvers/solver.cpp
